+++ src/sampler.h
@@ -45,7 +45,9 @@ struct perfidious_sampler_snapshot
     uint64_t values[PERFIDIOUS_METRIC_COUNT];
 };
 
-PERFIDIOUS_LOCAL uint32_t perfidious_platform_sampler_supported_metrics(enum perfidious_scope_id scope);
+PERFIDIOUS_LOCAL zend_result perfidious_platform_sampler_supported_metrics(
+    uint32_t requested_metrics, enum perfidious_scope_id scope, uint32_t *supported_metrics
+);
 
 PERFIDIOUS_LOCAL zend_result perfidious_platform_sampler_open(
     uint32_t metrics, enum perfidious_scope_id scope, struct perfidious_platform_sampler **sampler